requirements.txt
blinker==1.4
Bootstrap-Flask==1.5.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==1.16.0
firebase==3.0.1
Flask==1.1.2
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
gcloud==0.18.3
googleapis-common-protos==1.53.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==3.0.0
phonenumbers==8.12.19
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pymongo==3.11.4
pyparsing==2.4.7
python-jwt==2.0.1
pytz==2021.1
PyYAML==5.4.1
razorpay==1.2.0
requests==2.25.1
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
